requirements.txt
boto3
python-dotenv
pyyaml
sb-cli    